split the setup wizard into src/ui/credentials/
